created a resendVerification hook, to avoid code repetion
