+++ src/components/ActionBanner/ActionBanner.jsx
@@ -3,11 +3,11 @@ import styles from './ActionBanner.module.css';
 import { TestId } from './constants';
 import Button from '../Button';
 
-export const ActionBanner = ({ buttonLabel, className, dataTest, handleCTAClick, text, type }) => {
+export const ActionBanner = ({ buttonLabel, className, dataTest, handleCTAClick, text, type, isLoading }) => {
   return (
     <aside role="alert" className={[`${styles[`banner--${type}`]} ${styles['banner']}`, `${className}`].join(' ')} data-testid={dataTest}>
       <p className={styles.bannerText}>{text}</p>
-      <Button label={buttonLabel} size="medium" onClick={handleCTAClick} />
+      <Button label={buttonLabel} size="medium" onClick={handleCTAClick} disabled={isLoading} />
     </aside>
   );
 };
@@ -18,7 +18,8 @@ ActionBanner.propTypes = {
   dataTest: PropTypes.string,
   handleCTAClick: PropTypes.func,
   text: PropTypes.string,
-  type: PropTypes.oneOf(['sticky', 'default'])
+  type: PropTypes.oneOf(['sticky', 'default']),
+  isLoading: PropTypes.bool
 };
 
 ActionBanner.defaultProps = {
@@ -27,5 +28,6 @@ ActionBanner.defaultProps = {
   dataTest: TestId.ACTION_BANNER_TEST_ID,
   handleCTAClick: () => {},
   text: 'Some text',
-  type: 'sticky'
+  type: 'sticky',
+  isLoading: false
 };
