+++ src/sections/LearnerDashboard/Home/Home.jsx
@@ -9,6 +9,7 @@ import { APPLICATION_HISTORY, authorizedPost } from '../../../services/ApiClient
 import { useSelector } from 'react-redux';
 import ActionBanner from '../../../components/ActionBanner';
 import { PageCopy } from './constants';
+import useResendVerification from '../../../hooks/useResendVerification';
 const { dashboard } = constants;
 
 export const Home = () => {
@@ -18,16 +19,12 @@ export const Home = () => {
   const [singleEntry, setSingleEntry] = useState([]);
   const [data, setData] = useState([]);
   const { name, userId, isVerified } = useSelector((state) => state.user);
+  const { handleResendVerification, isLoading } = useResendVerification();
   const [first] = name.split(' ');
 
   const handleOneClick = (id) => {
     setApplicationTable(!applicationTable);
     setSingleEntry(data?.filter((entry) => entry._id === id));
-  };
-
-  const handleVerifyClick = () => {
-    // TODO: Make call to verify user
-    console.log('Verify user called');
   };
 
   useEffect(() => {
@@ -74,10 +71,11 @@ export const Home = () => {
       ) : (
         <ActionBanner
           dataTest={TestId.VERIFY_BANNER}
-          buttonLabel={PageCopy.VERIFY_BUTTON_LABEL}
+          buttonLabel={isLoading ? PageCopy.VERIFYING : PageCopy.VERIFY_BUTTON_LABEL}
           className={styles.verifyBanner}
-          handleCTAClick={handleVerifyClick}
+          handleCTAClick={handleResendVerification}
           text={PageCopy.VERIFY_TEXT}
+          isLoading={isLoading}
         />
       )}
 
@@ -89,8 +87,7 @@ export const Home = () => {
                 <button
                   key={idx}
                   className={currentTable === idx ? `${styles.tab}` : `${styles.tab} ${styles.tabAlt}`}
-                  onClick={() => setCurrentTable(idx)}
-                >
+                  onClick={() => setCurrentTable(idx)}>
                   {itm}
                 </button>
               );
