+++ src/sections/LearnerDashboard/Home/constants.js
@@ -1,4 +1,5 @@
 export const PageCopy = {
   VERIFY_BUTTON_LABEL: 'Verify account',
+  VERIFYING: 'Verifying...',
   VERIFY_TEXT: 'Your application has been received but your account is not yet verified. Click on the button below to verify your account.'
 };
